pyproject.toml
[tool.poetry]
name = "diana"
version = "0.1.0"
description = "diana"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_alipayvoice", "nonebot_plugin_hitokoto", "nonebot_plugin_answersbook", "nonebot_plugin_rtfm", "haruka_bot", "nonebot_plugin_boardgame", "nonebot_plugin_baidutranslate", "nonebot_plugin_caiyunai", "nonebot_plugin_covid19_news", "nonebot_plugin_admin"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
